+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added support for `enableGPUFrameCaptureMode` #1251 @bsudekum
 - Config setting presets can now also be loaded from the main bundle when bundling XcodeGenKit #1135 @SofteqDG
+- Added support for project files, files that don't belong to any target #1268 @skofgar
 
 ### Fixed